# Make aegir run faster