--- a/bin/build
+++ b/bin/build
@@ -6,7 +6,7 @@ const gulp = require('gulp')
 const args = require('args-parser')(process.argv)
 
 require('../src/gulp-log')(gulp)
-require('../gulp')(gulp)
+require('../gulp')(gulp, ['build'])
 
 if (args.browser) {
   gulp.start('build:browser')