--- a/bin/coverage
+++ b/bin/coverage
@@ -9,7 +9,7 @@ const path = require('path')
 const args = require('args-parser')(process.argv)
 
 require('../src/gulp-log')(gulp)
-require('../gulp')(gulp)
+require('../gulp')(gulp, ['coverage'])
 
 if (args.publish) {
   // ./node_modules/.bin/coveralls --verbose < coverage/lcov.info