--- a/bin/test
+++ b/bin/test
@@ -6,7 +6,7 @@ const gulp = require('gulp')
 const args = require('args-parser')(process.argv)
 
 require('../src/gulp-log')(gulp)
-require('../gulp')(gulp)
+require('../gulp')(gulp, ['test'])
 
 if (args.browser) {
   gulp.start('test:browser')