--- a/package.json
+++ b/package.json
@@ -33,7 +33,7 @@
     "chalk": "^1.1.3",
     "conventional-github-releaser": "^1.1.3",
     "coveralls": "^2.11.14",
-    "eslint": "^3.5.0",
+    "eslint": "^3.6.0",
     "eslint-config-standard": "^6.0.1",
     "eslint-plugin-promise": "^2.0.1",
     "eslint-plugin-standard": "^2.0.0",