--- a/package.json
+++ b/package.json
@@ -63,7 +63,7 @@
     "karma-sourcemap-loader": "^0.3.7",
     "karma-webpack": "^1.8.0",
     "lodash.camelcase": "^4.3.0",
-    "lodash.contains": "^2.4.3",
+    "lodash.includes": "^4.3.0",
     "lodash.upperfirst": "^4.3.1",
     "mocha": "^3.0.2",
     "phantomjs-prebuilt": "^2.1.12",