--- a/tasks/build.js
+++ b/tasks/build.js
@@ -1,5 +1,8 @@
 'use strict'
 
-module.exports = {
-  dep: ['build:browser', 'build:node']
+module.exports = (gulp) => {
+  require('./build/node')(gulp)
+  require('./build/browser')(gulp)
+
+  gulp.task('build', ['build:browser', 'build:node'])
 }