--- a/tasks/build.js
+++ b/tasks/build.js
@@ -3,6 +3,7 @@
 module.exports = (gulp) => {
   require('./build/node')(gulp)
   require('./build/browser')(gulp)
+  require('./clean')(gulp)
 
   gulp.task('build', ['build:browser', 'build:node'])
 }