--- a/tasks/clean.js
+++ b/tasks/clean.js
@@ -1,0 +1,6 @@
+'use strict'
+
+module.exports = (gulp) => {
+  require('./clean/node')(gulp)
+  require('./clean/browser')(gulp)
+}