--- a/tasks/clean/node.js
+++ b/tasks/clean/node.js
@@ -2,6 +2,8 @@
 
 const rimraf = require('rimraf')
 
-module.exports = (gulp, done) => {
-  rimraf('./lib', done)
+module.exports = (gulp) => {
+  gulp.task('clean:node', (done) => {
+    rimraf('./lib', done)
+  })
 }