--- a/tasks/lint.js
+++ b/tasks/lint.js
@@ -1,19 +1,20 @@
 'use strict'
 
-const $ = require('gulp-load-plugins')()
-const path = require('path')
+module.exports = (gulp) => {
+  gulp.task('lint', () => {
+    const eslint = require('gulp-eslint')
 
-module.exports = (gulp, done) => {
-  return gulp.src([
-    '*.js',
-    'bin/**',
-    'config/**/*.js',
-    'test/**/*.js',
-    'src/**/*.js',
-    'tasks/**/*.js',
-    'examples/**/*.js'
-  ])
-    .pipe($.eslint(path.resolve(__dirname, '../config/eslintrc.yml')))
-    .pipe($.eslint.format())
-    .pipe($.eslint.failAfterError())
+    return gulp.src([
+      '*.js',
+      'bin/**',
+      'config/**/*.js',
+      'test/**/*.js',
+      'src/**/*.js',
+      'tasks/**/*.js',
+      'examples/**/*.js'
+    ])
+      .pipe(eslint(`${__dirname}/../config/eslintrc.yml`))
+      .pipe(eslint.format())
+      .pipe(eslint.failAfterError())
+  })
 }