--- a/tasks/release.js
+++ b/tasks/release.js
@@ -1,21 +1,39 @@
 'use strict'
 
-const $ = require('gulp-load-plugins')()
-const runSequence = require('run-sequence')
-const _ = require('lodash')
+const contains = require('lodash.contains')
 
-function getEnv () {
-  if (_.includes($.util.env._, 'browser')) return 'browser'
-  if (_.includes($.util.env._, 'node')) return 'node'
-  if (_.includes($.util.env._, 'no-build')) return 'no-build'
-  if ($.util.env.env) return $.util.env.env
+function getEnv (env) {
+  if (contains(env._, 'browser')) return 'browser'
+  if (contains(env._, 'node')) return 'node'
+  if (contains(env._, 'no-build')) return 'no-build'
+  if (env.env) return env.env
 
   return 'default'
 }
 
-module.exports = (gulp, done) => {
-  runSequence.use(gulp)(
-    `release:${getEnv()}`,
-    done
-  )
+module.exports = (gulp) => {
+  require('./release/node')(gulp)
+  require('./release/browser')(gulp)
+  require('./release/no-build')(gulp)
+
+  require('./release/bump')(gulp)
+  require('./release/changelog')(gulp)
+  require('./release/commit')(gulp)
+  require('./release/contributors')(gulp)
+  require('./release/default')(gulp)
+  require('./release/github')(gulp)
+  require('./release/post-build')(gulp)
+  require('./release/pre-build')(gulp)
+  require('./release/publish')(gulp)
+  require('./release/push')(gulp)
+
+  gulp.task('release', (done) => {
+    const util = require('gulp-util')
+    const runSequence = require('run-sequence')
+
+    runSequence.use(gulp)(
+      `release:${getEnv(util.env)}`,
+      done
+    )
+  })
 }