--- a/tasks/release/contributors.js
+++ b/tasks/release/contributors.js
@@ -1,12 +1,14 @@
 'use strict'
 
-const $ = require('gulp-load-plugins')()
-const contributors = require('../../src/gulp-contributors')
+module.exports = (gulp) => {
+  gulp.task('release:contributors', () => {
+    const git = require('gulp-git')
+    const contributors = require('../../src/gulp-contributors')
 
-module.exports = (gulp, done) => {
-  return gulp.src('./package.json')
-    .pipe(contributors())
-    .pipe(gulp.dest('./'))
-    .pipe($.git.add())
-    .pipe($.git.commit('chore: update contributors', {args: '-n --allow-empty'}))
+    return gulp.src('./package.json')
+      .pipe(contributors())
+      .pipe(gulp.dest('./'))
+      .pipe(git.add())
+      .pipe(git.commit('chore: update contributors', {args: '-n --allow-empty'}))
+  })
 }