--- a/tasks/release/default.js
+++ b/tasks/release/default.js
@@ -1,12 +1,14 @@
 'use strict'
 
-const runSequence = require('run-sequence')
+module.exports = (gulp) => {
+  gulp.task('release:default', (done) => {
+    const runSequence = require('run-sequence')
 
-module.exports = (gulp, done) => {
-  runSequence.use(gulp)(
-    'release:pre-build',
-    'build',
-    'release:post-build',
-    done
-  )
+    runSequence.use(gulp)(
+      'release:pre-build',
+      'build',
+      'release:post-build',
+      done
+    )
+  })
 }