--- a/tasks/release/post-build.js
+++ b/tasks/release/post-build.js
@@ -1,16 +1,18 @@
 'use strict'
 
-const runSequence = require('run-sequence')
+module.exports = (gulp) => {
+  gulp.task('release:post-build', (done) => {
+    const runSequence = require('run-sequence')
 
-module.exports = (gulp, done) => {
-  runSequence.use(gulp)(
-    'release:contributors',
-    'release:bump',
-    'release:changelog',
-    'release:commit',
-    'release:push',
-    'release:github',
-    'release:publish',
-    done
-  )
+    runSequence.use(gulp)(
+      'release:contributors',
+      'release:bump',
+      'release:changelog',
+      'release:commit',
+      'release:push',
+      'release:github',
+      'release:publish',
+      done
+    )
+  })
 }