--- a/tasks/release/push.js
+++ b/tasks/release/push.js
@@ -1,16 +1,19 @@
 'use strict'
 
-const $ = require('gulp-load-plugins')()
+module.exports = (gulp) => {
+  gulp.task('release:push', (done) => {
+    const util = require('gulp-util')
+    const git = require('gulp-git')
 
-const utils = require('../../src/utils')
+    const utils = require('../../src/utils')
 
-module.exports = (gulp, done) => {
-  const remote = $.util.remote || 'origin'
-  $.util.log('Pushing to git...')
-  $.git.push(remote, 'master', {args: '--tags'}, (err) => {
-    if (err) return utils.fail(err.message)
+    const remote = util.remote || 'origin'
+    util.log('Pushing to git...')
+    git.push(remote, 'master', {args: '--tags'}, (err) => {
+      if (err) return utils.fail(err.message)
 
-    $.util.log(`Pushed to git ${remote}:master`)
-    done()
+      util.log(`Pushed to git ${remote}:master`)
+      done()
+    })
   })
 }