--- a/tasks/test.js
+++ b/tasks/test.js
@@ -1,15 +1,18 @@
 'use strict'
 
-const runSequence = require('run-sequence')
+module.exports = (gulp) => {
+  require('./test/node')(gulp)
+  require('./test/browser')(gulp)
 
-const utils = require('../src/utils')
+  gulp.task('test', (done) => {
+    const runSequence = require('run-sequence')
 
-module.exports = {
-  fn (gulp, done) {
+    const utils = require('../src/utils')
+
     runSequence.use(gulp)(
       'test:node',
       'test:browser',
       utils.exitOnFail(done)
     )
-  }
+  })
 }