--- a/tasks/test/node.js
+++ b/tasks/test/node.js
@@ -1,23 +1,24 @@
 'use strict'
 
-const $ = require('gulp-load-plugins')()
+module.exports = (gulp) => {
+  gulp.task('mocha', () => {
+    const spawnMocha = require('gulp-spawn-mocha')
 
-const utils = require('../../src/utils')
-const timeout = require('../../config/custom').timeout
+    const timeout = require('../../config/custom').timeout
 
-module.exports = {
-  fn (gulp, done) {
-    gulp.task('mocha', () => {
-      return gulp.src([
-        'test/node.js',
-        'test/**/*.spec.js'
-      ], {read: false})
-        .pipe($.spawnMocha({
-          R: 'spec',
-          timeout: timeout
-        }))
-    })
+    return gulp.src([
+      'test/node.js',
+      'test/**/*.spec.js'
+    ], {read: false})
+      .pipe(spawnMocha({
+        R: 'spec',
+        timeout: timeout
+      }))
+  })
+
+  gulp.task('test:node', (done) => {
+    const utils = require('../../src/utils')
 
     utils.hooksRun(gulp, 'test:node', ['mocha'], utils.exitOnFail(done))
-  }
+  })
 }