### Before
- `npm run test:node 3.70s user 0.59s system 89% cpu 4.784 total`

### After

-  `npm run test:node  2.11s user 0.21s system 108% cpu 2.146 total`

Ref #60