@diasdavid 2016-09-26
Nice! You can see the CPU load increasing as well.

LGTM!